backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.1
firebase-admin==6.6.0
pandas==2.3.3
numpy==2.1.2
scikit-learn==1.5.2
